# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API RESTful - NoteSwift",
    "version": "1.0.0",
...
    "license": {
      "name": "MIT License",
      "url": "https://github.com/ViniciusGR797/noteswift-api/blob/main/LICENSE"
    }
  },
  "components": {
    "securitySchemes": {
      "jwt": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Para autenticar, coloque o token JWT no formato **Bearer _token_**."
      }
    },
    "schemas": {
...
      "Note": {
        "type": "object",
        "required": [
          "_id",
          "title",
          "body",
          "style",
          "trashed",
          "update_at"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Identificador único da nota",
            "example": "6123456789abcdef01234567"
          },
          "title": {
            "type": "string",
            "description": "Título da nota",
            "example": "Título da nota"
          },
          "body": {
            "type": "string",
            "description": "Corpo da nota",
            "example": "Corpo da nota"
          },
          "style": {
            "type": "string",
            "description": "Estilo da nota",
            "example": "Estilo da nota"
          },
          "trashed": {
            "type": "boolean",
            "description": "Indica se a nota está na lixeira",
            "example": true
          },
          "deleted_date": {
            "type": "string",
            "description": "Data de exclusão da nota",
            "example": "2022-01-01 10:30:00"
          },
          "update_at": {
            "type": "string",
            "description": "Data de atualização da nota",
            "example": "2022-01-01 10:30:00"
          }
        }
      },
      "NoteCreate": {
        "type": "object",
        "required": [
          "title",
          "body",
          "style",
          "folder_id"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Título da nota",
            "example": "Título da nota"
          },
          "body": {
            "type": "string",
            "description": "Corpo da nota",
            "example": "Corpo da nota"
          },
          "style": {
            "type": "string",
            "description": "Estilo da nota",
            "example": "Estilo da nota"
          },
          "folder_id": {
            "type": "string",
            "description": "Identificador único da pasta, onde a anotação será armazenada",
            "example": "6123456789abcdef01234567"
          }
        }
      },
      "NoteUpdate": {
        "type": "object",
        "required": [
          "title",
          "body",
          "style"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Título da nota",
            "example": "Título da nota"
          },
          "body": {
            "type": "string",
            "description": "Corpo da nota",
            "example": "Corpo da nota"
          },
          "style": {
            "type": "string",
            "description": "Estilo da nota",
            "example": "Estilo da nota"
          }
        }
      },
      "NoteMove": {
        "type": "object",
        "required": [
          "folder_id"
        ],
        "properties": {
          "folder_id": {
            "type": "string",
            "description": "Identificador único da pasta, onde a anotação será armazenada",
            "example": "6123456789abcdef01234567"
          }
        }
      },
      "Bin": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Note"
        }
      },
      "NoteList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/NoteDownload"
        }
      },
      "NoteDownload": {
        "type": "object",
        "required": [
          "_id"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Identificador único da nota",
            "example": "6123456789abcdef01234567"
          }
        }
      },
...
      "MoveSuccess": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Mensagem de sucesso ao mover",
            "example": "Movida com sucesso"
          }
        }
      },
      "DeleteSuccess": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Mensagem de sucesso ao excluir",
            "example": "Excluído com sucesso"
          }
        }
      },
...
      "BadRequest": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Requisição inválida. Alguns parâmetros podem estar faltando ou serem inválidos",
            "example": "Alguns parâmetros podem estar faltando ou serem inválidos"
          }
        }
      },
      "Unauthorized": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Não autorizado. O token JWT não foi fornecido ou é inválido",
            "example": "O token JWT não foi fornecido ou é inválido"
          }
        }
      },
      "Forbidden": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Proibido. O usuário não tem permissão para acessar o recurso solicitado",
            "example": "Não tem permissão para acessar o recurso solicitado"
          }
        }
      },
      "NotFound": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Nenhum dado encontrado. O recurso solicitado não existe",
            "example": "Nenhum dado encontrado"
          }
        }
      },
      "InternalServerError": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Erro interno ao manipular dados no serviço",
            "example": "Erro interno do servidor"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "jwt": []
    }
  ],
  "paths": {
...
    "/notes/{note_id}": {
      "parameters": [
        {
          "name": "note_id",
          "in": "path",
          "required": true,
          "description": "ID da anotação que será deletada",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Visualiza anotação pelo _id",
        "description": "Retorna as informações da anotação pelo _id",
        "tags": [
          "Note"
        ],
        "operationId": "get_note_by_id",
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualiza anotação",
        "description": "Atualiza as informações da anotação desejada",
        "tags": [
          "Note"
        ],
        "operationId": "update_note",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteUpdate"
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove anotação",
        "description": "Apaga dados de anotação",
        "tags": [
          "Note"
        ],
        "operationId": "delete_note",
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/notes": {
      "post": {
        "summary": "Cria uma nova anotação",
        "description": "Cria uma nova anotação com base nos dados fornecidos no corpo da requisição",
        "tags": [
          "Note"
        ],
        "operationId": "create_note",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/notes/download": {
      "post": {
        "summary": "Cria um arquivo PDF de anotações",
        "description": "Cria um arquivo PDF de anotações com base nos dados fornecidos no corpo da requisição",
        "tags": [
          "Note"
        ],
        "operationId": "download_note",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteList"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bin"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/notes/moved/{note_id}": {
      "parameters": [
        {
          "name": "note_id",
          "in": "path",
          "required": true,
          "description": "ID da anotação que será movida",
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "summary": "Mover anotação",
        "description": "Alterar pasta em que se encontra a anotação",
        "tags": [
          "Note"
        ],
        "operationId": "move_note",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteMove"
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoveSuccess"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/notes/trash/{note_id}": {
      "parameters": [
        {
          "name": "note_id",
          "in": "path",
          "required": true,
          "description": "ID da anotação que será movida para lixeira",
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "summary": "Mover anotação para lixeira",
        "description": "Reciclar anotação para lixeira",
        "tags": [
          "Note"
        ],
        "operationId": "trash_note",
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/notes/restore/{note_id}": {
      "parameters": [
        {
          "name": "note_id",
          "in": "path",
          "required": true,
          "description": "ID da anotação que será recuperada da lixeira",
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "summary": "Recuperar anotação da lixeira",
        "description": "Recuperar anotação em que se encontra na lixeira",
        "tags": [
          "Note"
        ],
        "operationId": "restore_note",
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}